IOS-4934 Fix merge
