+++ TangemSwapping/Interfaces/ExpressWallet.swift
@@ -11,6 +11,11 @@ import Foundation
 public protocol ExpressWallet {
     var currency: ExpressCurrency { get }
     var address: String { get }
+
+    // Maybe will be deleted. We still deciding, How it will work
+    var decimalCount: Int { get }
+
+    func getBalance() async throws -> Decimal
 }
 
 public extension ExpressWallet {
@@ -24,5 +29,17 @@ public extension ExpressWallet {
 
     var isToken: Bool {
         contractAddress != ExpressConstants.coinContractAddress
+    }
+
+    // Maybe will be deleted. We still deciding, How it will work
+    func convertToWEI(value: Decimal) -> Decimal {
+        let decimalValue = pow(10, decimalCount)
+        return value * decimalValue
+    }
+
+    // Maybe will be deleted. We still deciding, How it will work
+    func convertFromWEI(value: Decimal) -> Decimal {
+        let decimalValue = pow(10, decimalCount)
+        return value / decimalValue
     }
 }
