--- TangemSwapping/Manager/ExpressManager/ExpressConstants.swift
@@ -1,13 +1,0 @@
-//
-//  ExpressConstants.swift
-//  TangemSwapping
-//
-//  Created by Sergey Balashov on 10.11.2023.
-//  Copyright © 2023 Tangem AG. All rights reserved.
-//
-
-import Foundation
-
-public enum ExpressConstants {
-    public static let coinContractAddress = "0"
-}
